refactor: extract TitledSection + SectionList to layout module
TitledSection is now a core layout component with a `contained` prop
for optional card background. Both modes have identical padding so
headings always align across a page. SectionList provides consistent
vertical spacing between sections.

ScrollShelf and PersonalStatsCard use TitledSection(contained) instead
of the previous Section + TitledSection composition.

Old imports via @/components/ui still work (re-exported).
