+++ web/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { TitledSection } from "./titled-section";
+export { SectionList } from "./section-list";
